requirements.txt
seleniumbase>=4.29.0
flask>=3.0.0
webdriver-manager>=4.0.1
jieba>=0.42.1
networkx>=3.0
scipy>=1.10.0
numpy>=1.24.0
